pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devops-openenv"
version = "0.1.0"
description = "OpenEnv benchmark for realistic SaaS incident triage."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "openai",
    "openenv-core>=0.2.0",
    "pydantic",
    "pyyaml",
    "uvicorn",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["env", "graders", "tasks", "server"]
